test(message_assembler/series): move sequence tracking tests to dedicated module
Refactor sequence tracking unit and property tests from inline module in
`series.rs` into a separate `tests.rs` file to improve code organization and
readability.
